Enviando exercícios

+++ sem1_questao1.c
@@ -12,23 +12,23 @@ acima de 200 reais = 10%.
 
 int main() {
   //declaração de variaveis
-  float valorCompra, valorFinal;
+  double valorCompra, valorFinal;
 
   //exibe mensagem solicitando o valor da compra
   printf("Insira o valor total da sua compra:\n");
 
   //lê o valor de entrada
-  scanf("%f", &valorCompra);
+  scanf("%lf", &valorCompra);
 
   //condicional
   if(valorCompra<=100){ //se entrada menor ou igual que 100
-    printf("\nR$ %.2f", valorCompra); //imprima entrada
+    printf("\nR$ %.2lf", valorCompra); //imprima entrada
   } if(valorCompra>100 && valorCompra<=200){ //se entrada maior que 100 e menor ou igual a 200
     valorFinal=valorCompra-(valorCompra*0.05); //calcula 5% desconto
-    printf("\nR$ %.2f", valorFinal); //imprime valor com desconto
+    printf("\nR$ %.2lf", valorFinal); //imprime valor com desconto
   } if(valorCompra>200){ //se entrada maior que 200
     valorFinal=valorCompra-(valorCompra*0.1); //calcula 10% desconto
-    printf("\nR$ %.2f", valorFinal); //imprime valor com desconto
+    printf("\nR$ %.2lf", valorFinal); //imprime valor com desconto
   }
 
   return 0;
